+++ src/components/pager/__internal__/pager-navigation-link.component.tsx
@@ -11,6 +11,8 @@ interface PagerNavigationLinkProps {
   currentPage: number;
   /** Count of all the pages  */
   pageCount: number;
+  /** If true, sets css property visibility: hidden on all disabled elements  */
+  hideDisabledElements?: boolean;
   /** Pagination page size */
   pageSize: number;
   /** onClick Callback function */
@@ -31,6 +33,7 @@ const PagerNavigationLink = ({
   type,
   currentPage,
   pageCount,
+  hideDisabledElements,
   pageSize,
   onClick,
   onPagination,
@@ -85,8 +88,11 @@ const PagerNavigationLink = ({
 
   const { text } = navLinkConfig[type];
 
+  const hideDisabledButtons = hideDisabledElements && disabled();
+
   return (
     <StyledPagerLink
+      hideDisabledButtons={hideDisabledButtons}
       data-element={`pager-link-${type}`}
       disabled={disabled()}
       onClick={