+++ src/components/vertical-menu/vertical-menu-full-screen.context.ts
@@ -1,0 +1,7 @@
+import React from "react";
+
+const VerticalMenuFullScreenContext = React.createContext({
+  isFullScreen: false,
+});
+
+export default VerticalMenuFullScreenContext;