+++ src/locales/locale.ts
@@ -109,6 +109,11 @@ interface Locale {
       close: () => string;
     };
   };
+  verticalMenuFullScreen: {
+    ariaLabels: {
+      close: () => string;
+    };
+  };
 }
 
 export default Locale;