+++ src/locales/pl-pl.ts
@@ -130,6 +130,11 @@ const plPL: Locale = {
       close: () => "Zamknij",
     },
   },
+  verticalMenuFullScreen: {
+    ariaLabels: {
+      close: () => "Zamknij",
+    },
+  },
 };
 
 export default plPL;